+++ src/app/core/hal/castle.service.ts
