+++ src/app/core/hal/index.ts
