+++ src/app/core/index.ts
@@ -2,7 +2,7 @@ export { CoreModule } from './core.module';
 
 // PUBLIC exports only here
 export { PlayerService, PlaybackMetadata, ValidationMetadata } from './audio';
-export { CastleService, CmsService, FeederService, HalDoc } from './cms';
+export { CastleService, CmsService, FeederService, HalDoc } from './hal';
 export { ModalService } from './modal';
 export { Upload, UploadService, UUID } from './upload';
 export { ToastrService } from './toastr';
